[Oztechan/CCC#3354] Move UMP logic into AdManager
