+++ android/core/ad/android-core-ad.gradle.kts
@@ -70,7 +70,6 @@ dependencies {
     }
 
     libs.android.apply {
-        implementation(activity)
         google.apply {
             DeviceFlavour.GOOGLE.implementation(googleAds)
             DeviceFlavour.GOOGLE.implementation(ump)
