+++ android/core/ad/src/main/kotlin/com/oztechan/ccc/android/core/ad/AdManager.kt
@@ -5,6 +5,8 @@ import android.content.Context
 
 interface AdManager {
 
+    fun initAds(activity: Activity)
+
     fun getBannerAd(
         context: Context,
         width: Int,
