+++ gradle/libs.versions.toml
@@ -10,7 +10,6 @@ glance = "1.0.0"
 material3 = "1.2.1"
 androidDesugaring = "2.0.4"
 androidMaterial = "1.11.0"
-activity = "1.9.0"
 composeActivity = "1.8.2"
 constraintLayout = "2.1.4"
 koinCore = "3.5.6"
@@ -72,7 +71,6 @@ common-konsist = { module = "com.lemonappdev:konsist", version.ref = "konsist" }
 android-androidDesugaring = { module = "com.android.tools:desugar_jdk_libs", version.ref = "androidDesugaring" }
 android-androidMaterial = { module = "com.google.android.material:material", version.ref = "androidMaterial" }
 android-material3 = { module = "androidx.compose.material3:material3", version.ref = "material3" }
-android-activity = { module = "androidx.activity:activity", version.ref = "activity" }
 android-constraintLayout = { module = "androidx.constraintlayout:constraintlayout", version.ref = "constraintLayout" }
 android-composeCompiler = { module = "androidx.compose.compiler:compiler", version.ref = "composeCompiler" } # not used dependency, workaround for receiving kotlin compose compiler updates by renovate: https://github.com/renovatebot/renovate/issues/18354
 android-composeToolingPreview = { module = "androidx.compose.ui:ui-tooling-preview", version.ref = "compose" }
